requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
anthropic>=0.69.0
python-pptx==1.0.2
pypdf==5.1.0
pymupdf==1.25.1
python-dotenv==1.0.1
Pillow==11.0.0
lxml==5.3.0
python-multipart==0.0.20
aiofiles==24.1.0
